Q3 Added Modular changes
Added add shooting functionality into its own script instead of having it spaced out in the PlayerController script
