+++ Assets/Scripts/Mechanics/Shooter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2ab87b78210ee4ea788cb5a0d4f2f5d6
